feat: Refactor shell modules / rearrange
